Patched scripts now run!
Patched scripts now run without crashing the ren'py engine, added protected areas, warning highlighting, and quote escaping. The quote escaping causes some problems with doubled quotes on some lines that should not have quotes. but this is easily fixable with some search and replacing. Also, will be some possible problems caused by escaping quotes that are already escaped, but that can be easily fixed in future versions (or just search and replace for "//").